Next phase completed.  Added extra processor for message queues.

+++ src/main/java/nicebank/AutomatedTeller.java
@@ -1,7 +1,5 @@
 package nicebank;
 
-import nicebank.implementations.ITeller;
-
 public class AutomatedTeller implements ITeller {
     private CashSlot cashSlot;
 
@@ -14,7 +12,7 @@ public class AutomatedTeller implements ITeller {
     }
 
     @Override
-    public void withdrawFrom(Account account, float amount) {
+    public void withdrawFrom(Account account, int amount) {
         System.out.println("[AutomatedTeller] > withdrawFrom() - first version of this method > otgher is in AtmUserInterface");
         account.debit(amount);
         cashSlot.dispense(amount);
